Reordered some rows in mapping tables.
This does NOT actually change the data at all!  These tables were
apparently created with no key defined, so the rows were in arbitrary
order -- but when I created and populated the tables in MySQL on
nyarumaa, the keys were defined correctly, and InnoDB ordered them by
key.  This is about what should happen anyway and the discrepancy adds
clutter when dumping corrections, so I'm just committing the new order.