Re-rip G/S/C egg moves.
Level up, TM, and tutor moves have already been ripped, so this should
be the last.

There are no changes (from what we had before) to Crystal, and only a
few additions to Gold/Silver.

Also, just to be safe, i checked that the egg moves in Silver are the
same as in Gold.

Thanks once again to UPC--it's easier to find something when you know
what you're looking for.