Fix whoosh result sorting.
Now results are sorted by is-this-your-language (times levenshtein
distance, if appropriate), then by rough class of result (Pokémon, then
moves, then abilities, etc.) and finally by name.

This fixes a couple issues:

- If both a foreign name and a local name matched a wildcard lookup,
  you'll see the local name.  Before, you'd see whichever happened to be
  first alphabetically.

- Wildcard results are more likely to have useful stuff at the top,
  rather than being dominated by foreign junk and names of obscure
  locations.

This also updates our usage of the whoosh API, which was old and busted
as of 2.0 or so.