Whoosh 2.5 performance quickfix
In Whoosh 2.5, fields that can be sorted on need to specify
sortable=True or else take a large performance hit because whoosh
completely dumped its old method of speeding up sorting (automatic
caching) in favor of this brand new one.

These probably don't *all* need to be sortable, but hey.

Note that we now require Whoosh>=2.5 because the sortable keyword arg
didn't exist in exist in earlier versions.