Add unreleased moves.
I also changed Celebrate's effect to "unknown", which I meant to commit
separately but it got caught up in this one.  It has a unique effect ID,
and we don't know for sure that it does nothing.  Also, Splash's effect
(which it was sharing) says that the move is disabled by Gravity, which
isn't true of Celebrate.