Rerip nature names from X.
In previous gens, Spanish nature names only ever appeared in the
feminine form, so that's what we had.  Now it looks like the isolated
nature names default to masculine — they're still feminine on the status
screen, but in the box you get "Naturaleza:   [adjective]" and I guess
the adjective is considered to be on its own rather than modifying
naturaleza.

Italian doesn't follow suit so I guess it gets to be the one special
snowflake language where we don't have the default dictionary form for
natures.