Improve schema ergonomics a bit, especially around slicing
The previous approach was moving towards having each attribute on a
locus do multiple different things, depending on context, and I think
that was headed towards being a mess.

This idea is to have actual locus objects be dumb containers, and have
various wrappers that call methods on the attributes to do interesting
work.