Exclude fossil pokemon for now
Fossil pokémon don't fit with the rest of the gift pokemon: rather than
being given a pokémon directly, you get an item which can be exchanged
for a pokémon somewhere else ("revived"). To handle them properly we'd
have to add an item location table. I don't want to figure that out
right now, so just ignore the fossils for now.