Travis CI: check that dump is idempotent
Run git-diff after dumping the database to make sure that there are no
changes to the csv files.

We often get PRs from people who edit stuff by hand and don't know to do
a load/dump of the database afterwards. Usually this means they were
adding game data by hand, which we want to discourage, but even if it's
done for valid reasons it's good to canonicalize the csv files as soon
as possible to avoid interfering with future commits.