Travis CI: fix some config warnings
`sudo: false` used to be the way to opt in to their container-based
infrastructure. That system was retired in 2018 and so the `sudo` option
no longer does anything. Remove it.

Travi CI supports multiple OSes now. Add `os: linux` to be explicit
about which OS we want to test on.