workflows: Add github actions file
Should be basically the same as the Travis CI file except more verbose
(ugh) and a different version matrix.

- 2.7 3.4 3.5 3.6 3.7 pypy pypy3
+ 2.7 3.5 3.7 3.9 pypy3

Kept 2.7 because we still run it. Dropped pypy2 because who cares?

Dropped 3.4 because it's way old at this point. Added 3.9 because it's
the latest release. Kept 3.5 and 3.7 and dropped the even releases
because it seems silly to test five different releases - the oldest and
newest we support should be adequate (plus 3.7 because that's what i
have.)